REPO/SCRIPTS: clean-up
* fixed line endings
* removed extra line spacing
*fix trialling whitespaces

DEVNOTE:
 EasternKingdoms scripting is almost completely in place...
still to be finished, is removal of SD2 kindship... don't understand why TC
had let this back in... or didnt clear the rest of it out. so alot of codesytling needs to change... mostly for azeroth... EventAI(EAI) needs xferred to SmartAI(SAI) for these zones. and also a brisk combing of the areas to make sure there all tied intogether as should be... You scriptors out there, EK is ready for yer touchs. ;)