DB: Old dir
* Returned the old dir back to  its rightful place.
only when there is an update pack does the sql fixes from updates get moved here. they do not get deleted...

this is so we can look back and see whats been changed, incase mistakes were made, and for those that do not wish to do a full update pack import if they only need a few updates.