Merged aura_322_333_work to master.
these spells tested working with the new aura fixes...
minor adjustments may still need made.

DevNote: these spells to spelleffects "can" be considered hacky to some,
the right way is to move them to spellscripts and db... when we have someone to to that properly, and these spells are working, it will get done.