Core/Auth revert most of the changes from commit f2f8970c273c9dd0365e02170b3078aa5e37eb9d * made couple slight adjustments.
this was not a good fix. nothing changed but dropping our name.
We are but a Suburb of TrinityCore City (lol at analogy),
with this it entitles us to bare our own name and signature on things within this project, as well to have our own set of rules and policies.
ONE of which approval must be given b4 trinity ports...

Signed-off-by: Bootz <bootz@projectskyfire.org>