Core/Currencies: Fixed currencies, and crashes related to instanced battle areas, and rewards.
To move forward we had to take a step backwards, meaning,
i had to back off the currency precision fixes, and go back to the old SF core,
and redo everything, and adjust it for our current core.
currencies was not done while in rebase, so the back step was a necessity
to get work already previously done into current core, so we can from there update as needed.
now let us progress.. ;)
Thanks to all those that worked on this in the older core.

Signed-off-by: Bootz <bootz@projectskyfire.org>