Restructuring SMSG_SET_PHASE_SHIFT these two parts in 4.2x(+) patches are merged, for PhaseShifts/TerrainSwaps.
in 406a and 410 seems this was not the case, ofc with only 2 true maps to switch between one in Gilneas and one in Kezan
this was roughly handled. Sortah still feeling out the way this was done officially.
this is Wip so all this can change at moments notice. (at my pace doubtful but expect it still) :/