Fix all unacceptable Quests
This is the actual Error, of why you cannot accept hundreds of Quests. Worldserver.conf overrides function to auto accept Quests, so QuestFlag 8 and SpecialFlag 4 is depraced, since those Flags handle auto accept. Dunno why there where actually 100's of Quests in that Line, but in blizzlike no Quest is auto accepted, as in database wrongly explained. 

So this smart Query will override like a few 100 row's, to make Quests blizzlike and acceptable.